requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0
pyhwp>=0.1b15
olefile>=0.47
requests>=2.31.0
